refactor(session): extract the slots, ground and view clusters
Three more behavior clusters leave DesignSession so the file clears the
1,500-line bar the seam is measured against (1,619 -> 1,478):

- useSolverSlots: the A/B/C slot configs, the derived active-slot view, the
  three gear-modal mutators and the PyNEC-remap effect. backendTouchedRef is
  returned raw for the two "user picked a backend" JSX call sites.
- useGroundConfig: the ground/terrain state plus the derived ground_model,
  terrainKey and tab summary. No effects at all.
- useViewState: cut angles, view + camera projection, the canvas display
  toggles, the per-example camera reset and the arrow-key view cycler.

All bodies and both literal dep arrays are verbatim moves. useSolverSlots and
useGroundConfig add no effect motion; useViewState is called at the cut-angle
states' old position, which keeps the camera-reset effect in place and moves
the arrow-key listener up three slots, ahead of the linked-design-freq and
mobile-carousel effects it shares no state with.

The remaining clusters were left alone deliberately: the design/measurement
frequency cluster cannot move because its band-snap effect would then run
before the linked-design-freq effect and win the opposite setDesignFreq race,
and buildRequest / controlsRef / the solve and preview effects are the
ordering-critical core the seam design keeps in the component.
